Egzaminy posiadają pytania wielokrotnego wyboru (tylko admin)
Poprawki do Usera przygotowywujące forma pod kontroller zapisu rezultatu

+++ Modules/Exam/Http/Controllers/ExamController.php
@@ -86,6 +86,7 @@ class ExamController extends Controller
     public function saveUsers(SaveUsers $request){
         $UserBelongs = ExamUser::select('*')->where('exam_id',$request->testName)->get();
         $counterBack = 0;
+        if ($request->user != null) {
             foreach ($request->user as $i => $item) {
                 if ($item['set'] == 1 && !isset($item['check'])) {
                     ExamUser::where('user_id', $i)->where('exam_id', $request->testName)->first()->delete();
@@ -110,27 +111,29 @@ class ExamController extends Controller
                     }
                 }
             }
+        }
+        if ($request->group != null) {
+            foreach ($request->group as $i => $item) {
+                if ($item['set'] == 1 && !isset($item['check'])) {
+                    ExamUser::where('exam_id', $request->testName)->where('group_id', $i)->first()->delete();
+                }
 
-        foreach ($request->group as $i => $item) {
-            if ($item['set'] == 1 && !isset($item['check'])) {
-                ExamUser::where('exam_id', $request->testName)->where('group_id', $i)->first()->delete();
-            }
-
-            if (isset($item['check'])) {
-                if ($item['set'] == 0 && $item['check'] == 'on') {
-                    $counter = 0;
-                    foreach ($UserBelongs as $user) {
+                if (isset($item['check'])) {
+                    if ($item['set'] == 0 && $item['check'] == 'on') {
                         $counter = 0;
-                        if ($user->group_id == $i) {
-                            $counter = 1;
+                        foreach ($UserBelongs as $user) {
+                            $counter = 0;
+                            if ($user->group_id == $i) {
+                                $counter = 1;
+                            }
                         }
-                    }
-                    if ($counter == 0) {
-                        $belong = new ExamUser();
-                        $belong->exam_id = $request->testName;
-                        $belong->group_id = $i;
-                        $belong->save();
-                        $counterBack++;
+                        if ($counter == 0) {
+                            $belong = new ExamUser();
+                            $belong->exam_id = $request->testName;
+                            $belong->group_id = $i;
+                            $belong->save();
+                            $counterBack++;
+                        }
                     }
                 }
             }
